Full-App/hr-backend-final/requirements.txt
fastapi
pydantic
mangum
pymongo
langchain
boto3
python-dotenv
openai
pdfminer.six
langchain-community
python-docx
python-pptx
sse-starlette
opensearch-py
unicorn
requests-aws4auth
loguru
PyPDF2